// src/app/models/ContratoModel.ts
export class ContratoModel{
    id_contrato:number;
    id_cliente:number;
    id_plan:number;
    fecha_corte: Date;
    fecha_inicio: Date;
    dia_corte: number;
    estado: string;
    estatus: string;
    create_by: string;
    create_Date: Date;
    modify_by:string;
    modify_date: Date;
    val_install: number;
}