// src/app/models/UsuarioModel.ts
export class UsuarioModel {
    id_usuario: number;
    nombre: string;
    apellido: string;
    user_name: string;
    password: string;
    estado: string;
    usuario_creacion: string;
    fecha_creacion: Date;
    usuario_actualizacion: string;
    fecha_actualizacion: Date;
    id_perfil: number;
    id_empresa: number;
}